Check for sidecar/init based on istio-inject map, not parsing version tags
Problem: Istio vet checks for expected versions of sidecar & init
containers and emits warnings if it finds the wrong versions.  These
versions are based on trying to detect the istio version based on the
version tag of the container image for a deployment called "istio-mixer"
but the image version tag may not be the same thing as the istio
version.  For instance, some users might re-tag and push as
"my-private-registry/mixer:latest-to-use-in-prod".

Solution: Be more conservative for now.  The most common case is that
you've completed an Istio upgrade and want to know if any pods are still
running with old sidecars/init containers.  We can tell this by checking
if the images that would be injected for new pods is the same as the
images that are already injected.

For now, remove the other warning checks.  As the concept of "istio
version" gets better we can add enhanced versions of these checks back
in.

I tested by editing one of our test deployments and generated these warnings:

    Mismatched sidecar image - doomsday-clock-v1-411741385-l8b3b
    ============================================================
    WARNING: The pod doomsday-clock-v1-411741385-l8b3b in namespace default is
    running with sidecar proxy image quay.io/aspenmesh/istio:proxy_debug-0.4.1-am1
    but your environment is injecting quay.io/aspenmesh/istio:proxy_debug-0.4.0-am1
    for new workloads. Consider upgrading the sidecar proxy in the pod.

    Mismatched istio-init image - doomsday-clock-v1-411741385-l8b3b
    ===============================================================
    WARNING: The pod doomsday-clock-v1-411741385-l8b3b in namespace default is
    running with istio-init image quay.io/aspenmesh/istio:proxy_init-0.4.1-am1 but
    your environment is injecting quay.io/aspenmesh/istio:proxy_init-0.4.0-am1 for
    new workloads. Consider upgrading the istio-init container in the pod.