Util functions check global mTLS status for v1.0
Problem: In Istio v1.0, a new resource (MeshPolicy) is used to enable/disable
global mTLS instead of the Istio Config. Currently only one vetter checks for
the status of global mTLS (the mtlsprobes vetter), but there may be more that
rely on it in the future. We are currently not handling the new MeshPolicy
resource, and we need to do that to comply with the changes in Istio v1.0.

Solution: Utility functions have been added to mtlsutils to take in MeshPolicies
and determine if global mTLS is enabled or disabled. These functions will be
used in the mtlsprobes vetter to update it to work with Istio v1.0.