Istio-Vet 1.0 Updates Problem: Istio changed the structure and options for policies in 1.0. Instead of a global policy, they now use a Mesh Policy (not the same object type as an Auth Policy), and have added several options for enabling or disabling mTLS in a Mesh or Auth policy. Code needed to be updated. In addition, the Details API needed to return 4 statuses of mTLS settings instead of a boolean of enabled or disabled since Istio now allows "Permissive" as an mTLS option.
Solution: LoadAuthPolicies will now take a mesh policy so that an AuthPolicy struct can ostensibly load all the policies for the cluster. Functions that used to return booleans now are wrappers that convert the mutli-option mTLS status for a resource into a boolean to preserve Vetter behavior. AuthPolicyIsMtls() uses new logic to determine the mTLS state for a resource based on Istio's new options.