Added more detail to svc port prefix vet note
Problem: the vet note for service port prefix does not list the unsupported
port prefixes that were found in the note body, which makes it difficult to
rectify the config.

Solution: the long version of the vet note is now more verbose. It contains a
list of the unsupported port prefixes that were found for a service, and the
docs for the vetter have been updated to contain the supported prefixes as of
v1.1.0.