Make conflictingvirtualservicehost merging aware.
According to the
docs (https://istio.io/docs/ops/best-practices/traffic-management/)
and issue #73, Pilot will merge virtual services that are bound to a
gateway if they have the same route and host as opposed to before
where that would be an error. Update this vetter to be aware of this
situation.