Check for conflicts in a 2 pass manner instead of 1.
In the 1 pass approach there was a lot of subtlety around ensuring
that conflicts were not being double counted and that rules for the
same route were still being compared. Instead, compare the rules for
the same route and for subroutes in different steps.

If this approach is prohibitive computationally (if trees are wide or,
less likely, deep), we can revert to the original approach.