Only report conflicts with virtual services with the same host.
Before, we were warning on the same host-gateway combination. However,
the documentation states that virtual services are merged when the
host is the same and bound to a gateway.

This commit also creates a new error when multiple virtual services
define the same host, but one uses the default/sidecar gateway.