Add constructors for vetters using Listers.
In some versions of Istio (especially prevelant in Istio 1.5 and
lower), the CRDs set "PreserveUnknownFields" to true. When using the
stock Istio client-go informers, this can prevent the informer caches
to synce when you start the informers.

We work around this by generating our own informer factories and
caches (which preserve the Listers from Istio's client-go).

Because of different package locations, our generated
informers do not satisfy the same interfaces as Istio's
client-go so we can't use the NewVetter function.

This commit creates a new function to add an alternate way of creating
a vetter using listers instead of a factory to create the vetters.